NatureEngineWebApp/requirements.txt
alabaster==0.7.12
alembic==1.0.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
Axelrod==4.3.0
Babel==2.6.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.3
docutils==0.14
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
imagesize==1.1.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
more-itertools==4.3.0
numpy==1.15.2
packaging==18.0
pandas==0.23.4
pluggy==0.7.1
prompt-toolkit==1.0.15
py==1.6.0
Pygments==2.3.0
pylint==2.1.1
pyparsing==2.2.2
pytest==3.8.2
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.20.0
rq==0.12.0
scipy==1.1.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.12
toolz==0.9.0
tqdm==4.26.0
typed-ast==1.1.0
urllib3==1.24.1
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1

#For heroku
gunicorn==19.7.1